backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
paho-mqtt==1.6.1
PyJWT==2.8.0
bcrypt==5.0.0
